Revert "Move saveTimer from RecordTimer to Timer"
This reverts commit 4565097de4205fea0efca43a3ead1206dfc65b76.
