+++ timer.py
@@ -1,16 +1,7 @@
-import datetime
-import os
-
 from bisect import insort
 from time import time, localtime, mktime
 from enigma import eTimer
-
-from Components.config import config
-from Tools import Directories
-from Tools.XMLTools import stringToXML
-
-import RecordTimer
-
+import datetime
 
 class TimerEntry:
 	StateWaiting  = 0
@@ -155,7 +146,6 @@ class Timer:
 	def __init__(self):
 		self.timer_list = [ ]
 		self.processed_timers = [ ]
-		self.Filename = Directories.resolveFilename(Directories.SCOPE_CONFIG, "timers.xml")
 
 		self.timer = eTimer()
 		self.timer.callback.append(self.calcNextActivation)
@@ -167,71 +157,6 @@ class Timer:
 	def stateChanged(self, entry):
 		for f in self.on_state_change:
 			f(entry)
-
-	def saveTimer(self):
-		list = []
-
-		list.append('<?xml version="1.0" ?>\n')
-		list.append('<timers>\n')
-
-		for timer in self.timer_list + self.processed_timers:
-			if timer.dontSave:
-				continue
-
-			list.append('<timer')
-			list.append(' begin="' + str(int(timer.begin)) + '"')
-			list.append(' end="' + str(int(timer.end)) + '"')
-			list.append(' serviceref="' + stringToXML(str(timer.service_ref)) + '"')
-			list.append(' repeated="' + str(int(timer.repeated)) + '"')
-			list.append(' name="' + str(stringToXML(timer.name)) + '"')
-			list.append(' description="' + str(stringToXML(timer.description)) + '"')
-			list.append(' afterevent="' + str(stringToXML({
-				RecordTimer.AFTEREVENT.NONE: "nothing",
-				RecordTimer.AFTEREVENT.STANDBY: "standby",
-				RecordTimer.AFTEREVENT.DEEPSTANDBY: "deepstandby",
-				RecordTimer.AFTEREVENT.AUTO: "auto"
-				}[timer.afterEvent])) + '"')
-			if timer.eit is not None:
-				list.append(' eit="' + str(timer.eit) + '"')
-			if timer.dirname:
-				list.append(' location="' + str(stringToXML(timer.dirname)) + '"')
-			if timer.tags:
-				list.append(' tags="' + str(stringToXML(' '.join(timer.tags))) + '"')
-			if timer.disabled:
-				list.append(' disabled="' + str(int(timer.disabled)) + '"')
-			list.append(' justplay="' + str(int(timer.justplay)) + '"')
-			list.append(' always_zap="' + str(int(timer.always_zap)) + '"')
-			list.append(' pipzap="' + str(int(timer.pipzap)) + '"')
-			list.append(' zap_wakeup="' + str(timer.zap_wakeup) + '"')
-			list.append(' rename_repeat="' + str(int(timer.rename_repeat)) + '"')
-			list.append(' conflict_detection="' + str(int(timer.conflict_detection)) + '"')
-			list.append(' descramble="' + str(int(timer.descramble)) + '"')
-			list.append(' record_ecm="' + str(int(timer.record_ecm)) + '"')
-			if timer.flags:
-				list.append(' flags="' + ' '.join([stringToXML(x) for x in timer.flags]) + '"')
-			list.append('>\n')
-
-			if config.recording.debug.value:
-				for time, code, msg in timer.log_entries:
-					list.append('<log')
-					list.append(' code="' + str(code) + '"')
-					list.append(' time="' + str(time) + '"')
-					list.append('>')
-					list.append(str(stringToXML(msg)))
-					list.append('</log>\n')
-
-			list.append('</timer>\n')
-
-		list.append('</timers>\n')
-
-		file = open(self.Filename + ".writing", "w")
-		for x in list:
-			file.write(x)
-		file.flush()
-
-		os.fsync(file.fileno())
-		file.close()
-		os.rename(self.Filename + ".writing", self.Filename)
 
 	def cleanup(self):
 		self.processed_timers = [entry for entry in self.processed_timers if entry.disabled]
